requirements.txt
fastapi==0.111.0
python-dotenv==1.0.1
google-generativeai==0.7.2
genai==1.0.0
pillow==10.4.0
PyJWT==2.8.0
pymongo==4.8.0
redis==5.0.7